and first and last indicator apply to

+++ lib/entity/candle_indicator_entity.dart
@@ -27,7 +27,7 @@ class CandleIndicatorEntity {
   LineStyle style = LineStyle.normal;
 
   String? color, secondColor, levelsColor;
-  List<int> levels = [];
+  List<double> levels = [];
 
   double value = 0;
   double? up;
