+++ lib/entity/indicator_entity.dart
@@ -47,7 +47,7 @@ class IndicatorEntity {
 
   String? color, secondColor, levelsColor;
 
-  List<int> levels = [];
+  List<double> levels = [];
   double value = 0;
   double? up;
   double? dn;
